requirements.txt
requests
aiohttp
tqdm
watchdog
python-dotenv
instaloader
asyncio
